+++ app/src/main/AndroidManifest.xml
@@ -59,7 +59,7 @@
             android:exported="true" />
 
         <service
-            android:name="com.apptec.registrateapp.repository.webservices.MyFirebaseMessagingService"
+            android:name="com.apptec.registrateapp.mainactivity.fnotification.CustomFirebaseMessagingService"
             android:exported="false">
             <intent-filter>
                 <action android:name="com.google.firebase.MESSAGING_EVENT" />
