+++ app/src/main/java/com/apptec/registrateapp/loginactivity/LoginInteractorImpl.java
@@ -10,6 +10,8 @@ import androidx.core.app.ActivityCompat;
 
 import com.apptec.registrateapp.App;
 import com.apptec.registrateapp.R;
+import com.apptec.registrateapp.mainactivity.fdevice.DeviceRetrofitInterface;
+import com.apptec.registrateapp.mainactivity.fpermission.PermissionTypeRetrofitInterface;
 import com.apptec.registrateapp.models.Company;
 import com.apptec.registrateapp.models.Device;
 import com.apptec.registrateapp.models.PermissionType;
@@ -21,9 +23,6 @@ import com.apptec.registrateapp.repository.localdatabase.DatabaseAdapter;
 import com.apptec.registrateapp.repository.localdatabase.RoomHelper;
 import com.apptec.registrateapp.repository.sharedpreferences.SharedPreferencesHelper;
 import com.apptec.registrateapp.repository.webservices.ApiClient;
-import com.apptec.registrateapp.repository.webservices.interfaces.DeviceRetrofitInterface;
-import com.apptec.registrateapp.repository.webservices.interfaces.LoginRetrofitInterface;
-import com.apptec.registrateapp.repository.webservices.interfaces.PermissionTypeRetrofitInterface;
 import com.apptec.registrateapp.repository.webservices.pojoresponse.loginresponse.LoginResponse;
 import com.apptec.registrateapp.util.Constants;
 import com.google.gson.JsonArray;
