+++ app/src/main/java/com/apptec/registrateapp/mainactivity/fpermission/PermissionListAdapter.java
@@ -14,7 +14,6 @@ import android.widget.TextView;
 import com.apptec.registrateapp.R;
 import com.apptec.registrateapp.models.Permission;
 import com.apptec.registrateapp.repository.webservices.ApiClient;
-import com.apptec.registrateapp.repository.webservices.interfaces.PermissionRetrofitInterface;
 import com.apptec.registrateapp.util.Constants;
 import com.google.gson.JsonObject;
